fix hot reload warning
